+++ CHANGELOG.md
@@ -241,6 +241,7 @@ realm.write(() => {
     ```
 ### Enhancements
 * Small improvement to performance by caching JSI property String object [#4863](https://github.com/realm/realm-js/pull/4863)
+* Add support for using functions as default property values, in order to allow dynamic defaults [#5001](https://github.com/realm/realm-js/pull/5001), [#2393](https://github.com/realm/realm-js/issues/2393)
 
 ### Compatibility
 * React Native >= v0.70.0